Moving CompositeModel into Model and fixed imports
