+++ eureka/S5_lightcurve_fitting/models/__init__.py
@@ -1,5 +1,4 @@
-from . import Model
-from .CompositeModel import CompositeModel
+from .Model import Model
 from .TransitModel import TransitModel
 from .PolynomialModel import PolynomialModel
 from .ExpRampModel import ExpRampModel
